add tailor and blacksmith
