+++ code/code/spec/spec_mobs.cc
@@ -4239,327 +4239,6 @@ int pestilence(TBeing *, cmdTypeT cmd, const char *, TMonster *me, TObj *)
   return FALSE;
 }
 
-static int engraveCost(TObj *obj, TBeing *ch, unsigned int shop_nr)
-{
-  double cost;
-
-  cost = obj->obj_flags.cost;
-
-  cost *= max(1,obj->obj_flags.cost/10000);
-
-  if(shop_nr)
-    cost *= shop_index[shop_nr].getProfitBuy(obj, ch);
-
-  return (int) cost;
-}
-
-int engraver(TBeing *ch, cmdTypeT cmd, const char *arg, TMonster *me, TObj *o)
-{
-  char buf[256];
-  TObj *item;
-  TBeing *final_pers;
-  int cost;
-  dirTypeT dir = DIR_NONE;
-  roomDirData *exitp;
-  int rc;
-  TBeing *tbt = NULL;
-  TThing *ttt;
-
-  class reg_struct {
-    public:
-    int wait;
-    int cost;
-    char *char_name;
-    char *obj_name;
-    reg_struct() :
-      wait(0),
-      cost(0),
-      char_name(NULL),
-      obj_name(NULL)
-    {
-    }
-    ~reg_struct() {
-      delete [] char_name;
-      char_name = NULL;
-      delete [] obj_name;
-      obj_name = NULL;
-    }
-  };
-  static reg_struct *job;
-
-  switch (cmd) {
-    case CMD_WHISPER:
-      return shopWhisper(ch, me, find_shop_nr(me->number), arg);
-    case CMD_GENERIC_DESTROYED:
-      delete (reg_struct *) me->act_ptr;
-      me->act_ptr = NULL;
-      return FALSE;
-    case CMD_GENERIC_CREATED:
-      if (!(me->act_ptr = new reg_struct())) {
-        perror("failed new of engraver.");
-        exit(0);
-      }
-      return FALSE;
-    case CMD_MOB_MOVED_INTO_ROOM:
-
-        return kick_mobs_from_shop(me, ch, (long int)o);
-
-    case CMD_MOB_VIOLENCE_PEACEFUL:
-      ttt = o;
-      tbt = dynamic_cast<TBeing *>(ttt);
-      me->doSay("Hey!  Take it outside.");
-      for (dir = MIN_DIR; dir < MAX_DIR; dir++) {
-        if (exit_ok(exitp = me->exitDir(dir), NULL)) {
-          act("$n throws you from $s shop.",
-                 FALSE, me, 0, ch, TO_VICT);
-          act("$n throws $N from $s shop.",
-                 FALSE, me, 0, ch, TO_NOTVICT);
-          me->throwChar(ch, dir, FALSE, SILENT_NO, true);
-          act("$n throws you from $s shop.",
-                 FALSE, me, 0, tbt, TO_VICT);
-          act("$n throws $N from $s shop.",
-                 FALSE, me, 0, tbt, TO_NOTVICT);
-          me->throwChar(tbt, dir, FALSE, SILENT_NO, true);
-          return TRUE;
-        }
-      }
-      return TRUE;
-    case CMD_GENERIC_PULSE:
-      if (!(job = (reg_struct *) me->act_ptr)) {
-        vlogf(LOG_PROC,"ENGRAVER PROC ERROR: terminating (hopefully)");
-        return FALSE;   
-      }
-      if (!job->char_name || !job->obj_name)
-        return FALSE;
-
-      if (job->wait > 0) {
-        job->wait--;
-        if (!job->wait) {
-          sprintf(buf, "That should do it %s!", job->char_name);
-          me->doSay(buf);
-          TThing *ts = NULL;
-          TObj *final = NULL;
-          if (!(ts = searchLinkedList(job->obj_name, me->stuff)) ||
-              !(final = dynamic_cast<TObj *>(ts))) {
-            me->doSay("Ack, I lost the item somehow! Tell a god immediately!  ");
-            vlogf(LOG_PROC,format("engraver lost his engraving item (%s)") %
-                (final ? final->name : ""));
-            return FALSE;
-          }
-          final->swapToStrung();
-
-          //  Remake the obj name.  
-          sprintf(buf, "%s %s", final->name.c_str(), job->char_name);
-          final->name = buf;
-
-          sprintf(buf, "This is the personalized object of %s", job->char_name);
-          final->action_description = buf;
-
-          if (!(final_pers = get_char_room(job->char_name, me->in_room))) {
-            me->doSay("Hmm, I seem to have lost the person I was engraving for.");
-            rc = me->doDrop(job->obj_name, NULL);
-            if (IS_SET_DELETE(rc, DELETE_THIS))
-              return DELETE_THIS;
-            return FALSE;
-          }
-          if ((final_pers->getCarriedVolume() + final->getTotalVolume())  > final_pers->carryVolumeLimit()) {
-            me->doSay("You can't carry it! I'll just drop it here for you!");
-            rc = me->doDrop(job->obj_name, NULL);
-            if (IS_SET_DELETE(rc, DELETE_THIS))
-              return DELETE_THIS;
-            delete [] job->char_name;
-            job->char_name = NULL;
-            delete [] job->obj_name;
-            job->obj_name = NULL;
-            job->cost = 0;
-            return FALSE;
-          }
-          // final-weight > free final carry weight
-          if (compareWeights(final->getTotalWeight(TRUE),
-                (final_pers->carryWeightLimit() - final_pers->getCarriedWeight())) == -1) {
-            me->doSay("You can't carry it! I'll just drop it here for you!");
-            rc = me->doDrop(job->obj_name, NULL);
-            if (IS_SET_DELETE(rc, DELETE_THIS))
-              return DELETE_THIS;
-            delete [] job->char_name;
-            job->char_name = NULL;
-            delete [] job->obj_name;
-            job->obj_name = NULL;
-
-            job->cost = 0;
-            return FALSE;
-          }
-          sprintf(buf, "%s %s", job->obj_name, job->char_name);
-          if (me->doGive(buf, GIVE_FLAG_IGN_DEX_TEXT) == DELETE_THIS)
-            return DELETE_THIS;
-          delete [] job->char_name;
-          job->char_name = NULL;
-          delete [] job->obj_name;
-          job->obj_name = NULL;
-
-          job->cost = 0;
-          return FALSE;
-        } else {
-          sprintf(buf, "$n etches part of %s's %s.", job->char_name, job->obj_name);
-          act(buf, FALSE, me, NULL, ch, TO_ROOM);
-          return FALSE;
-        }
-      }
-      break;
-    case CMD_NORTH:
-    case CMD_SOUTH:
-    case CMD_WEST:
-    case CMD_EAST:
-    case CMD_UP:
-    case CMD_DOWN:
-    case CMD_NE:
-    case CMD_SW:
-    case CMD_SE:
-    case CMD_NW:
-      if (!(job = (reg_struct *) me->act_ptr)) {
-        return FALSE;
-      }
-      if (!job->char_name) {
-        return FALSE;
-      }
-      if (job->char_name == ch->getName()) {
-        sprintf(buf, "%s! Don't leave until I finish with this %s!", ch->getName().c_str(), job->obj_name);
-        me->doSay(buf);
-        return TRUE;
-      } else
-        return FALSE;
-      break;
-    case CMD_VALUE: {
-      for(; *arg && isspace(*arg);arg++);
-      TThing *ts = NULL;
-      TObj *valued = NULL;
-      if (!(ts = searchLinkedListVis(ch, arg, ch->stuff)) ||
-          !(valued = dynamic_cast<TObj *>(ts))) {
-        me->doTell(ch->getName(), "You don't have that item.");
-        return TRUE;
-      }
-
-      if (valued->engraveMe(ch, me, false))
-        return TRUE;
-
-      if (valued->obj_flags.cost <=  500) {
-        me->doTell(ch->getName(), "This item is too cheap to be engraved.");
-        return TRUE;
-      }
-      if (!valued->action_description.empty()) {
-        me->doTell(ch->getName(), "This item has already been engraved!");
-        return TRUE;
-      }
-      if (obj_index[valued->getItemIndex()].max_exist <= 10) {
-        me->doTell(ch->getName(), "I refuse to engrave such an artifact of beauty!");
-        return TRUE;
-      }
-      if (valued->obj_flags.decay_time >= 0) {
-        me->doTell(ch->getName(), "Sorry, but this item won't last long enough to bother with an engraving!");
-        return TRUE;
-      }
-
-      cost = engraveCost(valued, ch, find_shop_nr(me->number));
-
-      me->doTell(ch->getName(), format("It will cost %d talens to engrave your %s.") % cost % fname(valued->name));
-      return TRUE;
-      }
-    case CMD_MOB_GIVEN_ITEM:
-      // prohibit polys and charms from engraving 
-      if (dynamic_cast<TMonster *>(ch)) {
-        me->doTell(fname(ch->name), "I don't engrave for beasts.");
-        return TRUE;
-      }
-      if (!(item = o)) {
-        me->doTell(ch->getName(), "You don't have that item!");
-        return TRUE;
-      }
-      me->logItem(item, CMD_EAST);  // log the receipt of the item
-
-      if (item->engraveMe(ch, me, true))
-        return TRUE;
-
-      if (item->obj_flags.cost <= 500) {
-        me->doTell(ch->getName(), "That can't be engraved!");
-        strcpy(buf, item->name.c_str());
-        strcpy(buf, add_bars(buf).c_str());
-        sprintf(buf + strlen(buf), " %s", fname(ch->name).c_str());
-        me->doGive(buf,GIVE_FLAG_IGN_DEX_TEXT);
-        return TRUE;
-      }
-      if (!item->action_description.empty()) {
-        me->doTell(ch->getName(), "Sorry, but this item has already been engraved!");
-        strcpy(buf, item->name.c_str());
-        strcpy(buf, add_bars(buf).c_str());
-        sprintf(buf + strlen(buf), " %s", fname(ch->name).c_str());
-        me->doGive(buf,GIVE_FLAG_IGN_DEX_TEXT);
-        return TRUE;
-      }
-      if (obj_index[item->getItemIndex()].max_exist <= 10) {
-        me->doTell(ch->getName(), "This artifact is too powerful to be engraved!");
-        strcpy(buf, item->name.c_str());
-        strcpy(buf, add_bars(buf).c_str());
-        sprintf(buf + strlen(buf), " %s", fname(ch->name).c_str());
-        me->doGive(buf,GIVE_FLAG_IGN_DEX_TEXT);
-        return TRUE;
-      }
-      if (item->obj_flags.decay_time >= 0) {
-        me->doTell(ch->getName(), "This won't be around long enough to bother engraving it!");
-        strcpy(buf, item->name.c_str());
-        strcpy(buf, add_bars(buf).c_str());
-        sprintf(buf + strlen(buf), " %s", fname(ch->name).c_str());
-        me->doGive(buf,GIVE_FLAG_IGN_DEX_TEXT);
-        return TRUE;
-      }
-
-      cost = engraveCost(item, ch, find_shop_nr(me->number));
-
-      if (ch->getMoney() < cost) {
-        me->doTell(ch->getName(), "I have to make a living! If you don't have the money, I don't do the work!");
-        strcpy(buf, item->name.c_str());
-        strcpy(buf, add_bars(buf).c_str());
-        sprintf(buf + strlen(buf), " %s", fname(ch->name).c_str());
-        me->doGive(buf,GIVE_FLAG_IGN_DEX_TEXT);
-        return TRUE;
-      }
-      job = (reg_struct *) me->act_ptr;
-      if (!job->wait || !job->char_name) {
-        job->wait = max(1, (int) (item->obj_flags.max_struct_points)/7);
-        sprintf(buf, "Thanks for your business, I'll take your %d talens payment in advance!", cost);
-        me->doSay(buf);
-
-	TShopOwned tso(find_shop_nr(me->number), me, ch);
-	tso.doBuyTransaction(cost, format("engraving %s") % item->getName(), 
-			     TX_BUYING_SERVICE);
-
-        job->cost = cost;
-        job->char_name = new char[ch->getName().length() + 1];
-        strcpy(job->char_name, ch->getName().c_str());
-        job->obj_name = new char[fname(item->name).length() + 1];
-        strcpy(job->obj_name, fname(item->name).c_str());
-        --(*item);
-        *me += *item; 
-
-	me->saveChar(Room::AUTO_RENT);
-	ch->saveChar(Room::AUTO_RENT);
-
-        return TRUE;
-      } else {
-        sprintf(buf, "Sorry, %s, but you'll have to wait while I engrave %s's item.", ch->getName().c_str(), job->char_name);
-        me->doSay(buf);
-        strcpy(buf, item->name.c_str());
-        strcpy(buf, add_bars(buf).c_str());
-        sprintf(buf + strlen(buf), " %s", fname(ch->name).c_str());
-        me->doGive(buf,GIVE_FLAG_IGN_DEX_TEXT);
-        return TRUE;
-      }
-      return FALSE;
-    default:
-      return FALSE;
-  }
-  return FALSE;
-}
-
 int fireMaster(TBeing *ch, cmdTypeT cmd, const char *arg, TMonster *me, TObj *)
 {
   char buf[256];
@@ -7504,6 +7183,8 @@ TMobSpecs mob_specials[NUM_MOB_SPECIALS + 1] =
   {FALSE, "pet veterinarian", petVeterinarian},
   {TRUE, "lagomorph", lagomorph},
   {TRUE, "Trainer: Offense", CDGenericTrainer},
+  {TRUE, "Tailor", tailor},
+  {TRUE, "Blacksmith", blacksmith},
 // replace non-zero, bogus_mob_procs above before adding
 };
