+++ code/code/spec/spec_mobs.h
@@ -215,7 +215,9 @@ const int SPEC_RATION_FACTORY     =217;
 const int SPEC_PET_VETERINARIAN   =218;
 const int SPEC_LAGOMORPH          =219;
 const int SPEC_TRAINER_OFFENSE    =220;
-const int NUM_MOB_SPECIALS        =220;
+const int SPEC_TAILOR             =221;
+const int SPEC_BLACKSMITH         =222;
+const int NUM_MOB_SPECIALS        =222;
 
 extern const int GET_MOB_SPE_INDEX(int d);
 
@@ -223,6 +225,13 @@ struct TMobSpecs {
   bool assignable;
   const char *name;
   int (*proc) (TBeing *, cmdTypeT, const char *, TMonster *, TObj *);
+};
+
+// Type for customization
+enum customizerType {
+    TYPE_ENGRAVER,
+    TYPE_TAILOR,
+    TYPE_BLACKSMITH
 };
 
 extern TMobSpecs mob_specials[NUM_MOB_SPECIALS + 1];
@@ -269,6 +278,10 @@ extern int heroFaerie(TBeing *ch, cmdTypeT cmd, const char *arg, TMonster *mysel
 extern int heroFaerie(TBeing *ch, cmdTypeT cmd, const char *arg, TMonster *myself, TObj *);
 extern int brickCollector(TBeing *, cmdTypeT, const char *, TMonster *, TObj *);
 extern int targetDummy(TBeing *, cmdTypeT, const char *, TMonster *, TObj *);
+extern int engraver(TBeing *ch, cmdTypeT cmd, const char *arg, TMonster *me, TObj *o);
+extern int tailor(TBeing *ch, cmdTypeT cmd, const char *arg, TMonster *me, TObj *o);
+extern int blacksmith(TBeing *ch, cmdTypeT cmd, const char *arg, TMonster *me, TObj *o);
+extern int customizer(TBeing *ch, cmdTypeT cmd, const char *arg, TMonster *me, TObj *o, customizerType custtype);
 
 class attune_struct {
   public:
